+++ src/OLD-API/Startup.cs
@@ -1,7 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.HttpsPolicy;
@@ -11,14 +7,23 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.OpenApi.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.EntityFrameworkCore;
+using lunch_project.Classes;
 
-namespace API
+namespace LunchAPI
 {
     public class Startup
     {
         public Startup(IConfiguration configuration)
         {
-            Configuration = configuration;
+            Configuration = new ConfigurationBuilder()
+                .AddConfiguration(configuration)
+                .AddEnvironmentVariables()
+                .Build();
         }
 
         public IConfiguration Configuration { get; }
@@ -26,11 +31,32 @@ namespace API
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
+            /*
+            services.AddDbContext<DataDbContext>(opt =>
+            {
+                opt.UseSqlServer("LunchAPI"); 
+            }); 
+            */
+            
+
+            /*
+            services.AddDbContext<DataDbContext>(options =>
+            {
+                options.UseSqlServer(Program.GetConnectionString());
+            });
+            */
+
+            /*
+            services.AddDbContext<DataDbContext>(options => 
+            { 
+                options.UseSqlServer(Configuration.GetConnectionString("LunchAPI")); 
+            });
+            */
 
             services.AddControllers();
             services.AddSwaggerGen(c =>
             {
-                c.SwaggerDoc("v1", new OpenApiInfo { Title = "API", Version = "v1" });
+                c.SwaggerDoc("v1", new OpenApiInfo { Title = "LunchAPI", Version = "v1" });
             });
         }
 
@@ -41,7 +67,7 @@ namespace API
             {
                 app.UseDeveloperExceptionPage();
                 app.UseSwagger();
-                app.UseSwaggerUI(c => c.SwaggerEndpoint("/swagger/v1/swagger.json", "API v1"));
+                app.UseSwaggerUI(c => c.SwaggerEndpoint("/swagger/v1/swagger.json", "LunchAPI v1"));
             }
 
             app.UseHttpsRedirection();
